+++ app/controllers/concerns/has_current_group.rb
@@ -1,22 +1,23 @@
 # SubdomainGroups
 #
-module SubdomainGroups
+module HasCurrentGroup
   extend ActiveSupport::Concern
 
   included do
-    expose(:current_group) { @current_group }
+    expose(:current_group) { load_subdomain_group }
   end
 
   protected
   def load_subdomain_group
     if request.subdomain.blank?
-      @current_group = Site.new
+      Site.new
     else
-      @current_group = Group.find_by_slug(request.subdomain)
-      if @current_group.blank?
+      current_group = Group.find_by_slug(request.subdomain)
+      if current_group.blank?
         puts "Subdomain #{request.subdomain}"
         raise ActionController::RoutingError.new('Not Found')
       end
+      current_group
     end
   end
 end