+++ app/controllers/profiles_controller.rb
@@ -1,6 +1,5 @@
 class ProfilesController < ApplicationController
   respond_to :html
-  before_filter :load_subdomain_group
 
   def edit
     respond_with current_group