+++ test/unit/group_test.rb
@@ -8,4 +8,9 @@ describe Group do
     m.user.must_equal group.user
     m.state.must_equal 'owner'
   end
+
+  it "has membership for user" do
+    group = create(:group)
+    group.membership_for(group.user).must_be :present?
+  end
 end